requirements.txt
Django==6.0.4
pdfplumber==0.11.9
PyPDF2==3.0.1
pytesseract==0.3.13
pillow==12.2.0
gunicorn==22.0.0
